signal_handler: avoid signed/unsigned mismatch (-Wsign-compare)
We're ignoring the result of this write() to stderr anyway, because
if it failed... what would we do? Write to stderr? That wouldn't work
any better the second time :-)

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=17289

+++ bus/main.c
@@ -87,7 +87,8 @@ signal_handler (int sig)
             static const char message[] =
               "Unable to write to reload pipe - buffer full?\n";
 
-            if (write (STDERR_FILENO, message, strlen (message)) != strlen (message))
+            if (write (STDERR_FILENO, message, strlen (message)) !=
+                (ssize_t) strlen (message))
               {
                 /* ignore failure to write out a warning */
               }
@@ -111,7 +112,8 @@ signal_handler (int sig)
               "Unable to write termination signal to pipe - buffer full?\n"
               "Will exit instead.\n";
 
-            if (write (STDERR_FILENO, message, strlen (message)) != strlen (message))
+            if (write (STDERR_FILENO, message, strlen (message)) !=
+                (ssize_t) strlen (message))
               {
                 /* ignore failure to write out a warning */
               }
